[FIX]#65191  Error message shows when using forms "Relocation" and "Hire new staff"

+++ dhHelpdeskNG/dhHelpdesk_NG.Service/Services/UniversalCase/UniversalCaseService.cs
@@ -918,6 +918,8 @@ namespace DH.Helpdesk.Services.Services.UniversalCase
             var conditionTypeId = (int)((ConditionType)Enum.Parse(typeof(ConditionType), ConditionType.MultiCaseSplit.ToString()));
 
             caseId = -1;
+            caseNumber = -1;
+
             var baseCaseId = -1;
             decimal caseNum = -1;
 
@@ -930,63 +932,70 @@ namespace DH.Helpdesk.Services.Services.UniversalCase
             //prevent email notifications from sending for base (multiform) case
             res = SaveCase(baseCaseModel, baseAuxModel, out baseCaseId, out caseNum, false);
 
-            try
+            if (res.IsSucceed)
             {
-                if (baseCaseSolution.SplitToCaseSolutionDescendants != null && baseCaseSolution.SplitToCaseSolutionDescendants.Any())
+                try
                 {
-                    var caseSolutions = baseCaseSolution.SplitToCaseSolutionDescendants
-                        .Where(x => x.SplitToCaseSolutionDescendant.Status > 0)
-                        .OrderBy(x => x.SplitToCaseSolutionDescendant.SortOrder);
-
-                    foreach (var splitToCaseSolutionEntity in caseSolutions)
+                    if (baseCaseSolution.SplitToCaseSolutionDescendants != null && baseCaseSolution.SplitToCaseSolutionDescendants.Any())
                     {
-                        var childCaseSolutionId = splitToCaseSolutionEntity.SplitToCaseSolutionDescendant.Id;
-                        var caseTemplate = _caseSolutionService.GetCaseSolution(childCaseSolutionId);
+                        var caseSolutions = baseCaseSolution.SplitToCaseSolutionDescendants
+                            .Where(x => x.SplitToCaseSolutionDescendant.Status > 0)
+                            .OrderBy(x => x.SplitToCaseSolutionDescendant.SortOrder);
 
-                        //TODO: refactor and make a long term solution, this is just for DepartmentId
-                        var departmentId = ChangeDepartmentId(baseCaseModel.Customer_Id, baseCaseModel.Department_Id, caseTemplate.Customer_Id);
-
-                        var doSplit = _conditionService.CheckConditions(baseCaseId, childCaseSolutionId, conditionTypeId);
-
-                        if (doSplit.Show)
+                        foreach (var splitToCaseSolutionEntity in caseSolutions)
                         {
-                            var descendantCaseModel = caseModel.DeepClone();
-                            descendantCaseModel.Department_Id = departmentId;
+                            var childCaseSolutionId = splitToCaseSolutionEntity.SplitToCaseSolutionDescendant.Id;
+                            var caseTemplate = _caseSolutionService.GetCaseSolution(childCaseSolutionId);
 
-                            //apply values from case solution
-                            descendantCaseModel = ApplyValuesFromCaseSolution(descendantCaseModel, childCaseSolutionId);
+                            //TODO: refactor and make a long term solution, this is just for DepartmentId
+                            var departmentId = ChangeDepartmentId(baseCaseModel.Customer_Id, baseCaseModel.Department_Id, caseTemplate.Customer_Id);
 
-                            // prevent extended case relation creation in SaveCase
-                            descendantCaseModel.ExtendedCaseData_Id = null;
-                            descendantCaseModel.ExtendedCaseForm_Id = null;
+                            var doSplit = _conditionService.CheckConditions(baseCaseId, childCaseSolutionId, conditionTypeId);
 
-                            int childCaseId;
-                            res = SaveCase(descendantCaseModel, baseAuxModel, out childCaseId, out caseNum);
-
-                            if (res.IsSucceed && childCaseId > 0)
+                            if (doSplit.Show)
                             {
-                                if (caseId == -1)
-                                    caseId = childCaseId;
+                                var descendantCaseModel = caseModel.DeepClone();
+                                descendantCaseModel.Department_Id = departmentId;
 
-                                if (baseCaseExtendedCaseDataId > 0)
+                                //apply values from case solution
+                                descendantCaseModel = ApplyValuesFromCaseSolution(descendantCaseModel, childCaseSolutionId);
+
+                                // prevent extended case relation creation in SaveCase
+                                descendantCaseModel.ExtendedCaseData_Id = null;
+                                descendantCaseModel.ExtendedCaseForm_Id = null;
+
+                                int childCaseId;
+                                res = SaveCase(descendantCaseModel, baseAuxModel, out childCaseId, out caseNum);
+
+                                if (res.IsSucceed && childCaseId > 0)
                                 {
-                                    var exCaseForm = caseTemplate.ExtendedCaseForms.FirstOrDefault();
-                                    if (exCaseForm != null)
-                                        _extendedCaseService.CopyExtendedCaseToCase(baseCaseExtendedCaseDataId, childCaseId, baseAuxModel.UserIdentityName, exCaseForm.Id);
+                                    if (caseId == -1)
+                                    {
+                                        caseNumber = caseNum;
+                                        caseId = childCaseId;
+                                    }
+
+                                    if (baseCaseExtendedCaseDataId > 0)
+                                    {
+                                        var exCaseForm = caseTemplate.ExtendedCaseForms.FirstOrDefault();
+                                        if (exCaseForm != null)
+                                        {
+                                            _extendedCaseService.CopyExtendedCaseToCase(baseCaseExtendedCaseDataId, childCaseId, baseAuxModel.UserIdentityName, exCaseForm.Id);
+                                        }
+                                    }
                                 }
                             }
                         }
                     }
                 }
-            }
-            finally
-            {
-                //delete base case
-                if (baseCaseId > 0)
-                    _caseService.Delete(baseCaseId, "", null);
+                finally
+                {
+                    //delete base case
+                    if (baseCaseId > 0)
+                        _caseService.Delete(baseCaseId, "", null);
+                }
             }
             
-            caseNumber = caseNum;
             return res;
         }
